requirements.txt
# Flask Web Framework
Flask>=3.0.0
Flask-SocketIO>=5.3.0

# Computer Vision & Image Processing
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0

# Google Gemini AI
google-genai>=1.0.0

# HTTP Requests
requests>=2.31.0

# OCR Engine (Fallback)
easyocr>=1.7.0

# Installation Notes:
# 1. Set GEMINI_API_KEY environment variable
# 2. EasyOCR used as fallback if Gemini fails